+++ windjview/PageIndexWnd.cpp
@@ -1,5 +1,5 @@
 //	WinDjView
-//	Copyright (C) 2004-2009 Andrew Zhezherun
+//	Copyright (C) 2004-2012 Andrew Zhezherun
 //
 //	This program is free software; you can redistribute it and/or modify
 //	it under the terms of the GNU General Public License as published by
@@ -15,8 +15,6 @@
 //	with this program; if not, write to the Free Software Foundation, Inc.,
 //	51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 //	http://www.gnu.org/copyleft/gpl.html
-
-// $Id$
 
 #include "stdafx.h"
 #include "WinDjView.h"
@@ -397,7 +395,7 @@ void CPageIndexWnd::OnChangeText()
 
 	text = MapCharacters(tolower(text));
 
-	int left = 0, right = m_sorted.size();
+	int left = 0, right = (int)m_sorted.size();
 	while (left < right - 1)
 	{
 		int mid = (left + right) / 2;
